+++ src/main/java/kr/co/conceptbe/common/entity/domain/MemberSkillCategory.java
@@ -1,14 +1,17 @@
-package kr.co.conceptbe.member;
+package kr.co.conceptbe.common.entity.domain;
 
 import static lombok.AccessLevel.PROTECTED;
 
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import kr.co.conceptbe.common.entity.domain.SkillCategory;
+import kr.co.conceptbe.member.Member;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -26,13 +29,23 @@ public class MemberSkillCategory {
     private Member member;
 
     @ManyToOne
-    @JoinColumn(name = "skill_category")
+    @JoinColumn(name = "skill_category_id")
     private SkillCategory skillCategory;
 
-    public MemberSkillCategory(Long id, Member member, SkillCategory skillCategory) {
+    @Column
+    @Enumerated(EnumType.STRING)
+    private SkillLevel skillLevel;
+
+    public MemberSkillCategory(
+            Long id,
+            Member member,
+            SkillCategory skillCategory,
+            SkillLevel skillLevel
+    ) {
         this.id = id;
         this.member = member;
         this.skillCategory = skillCategory;
+        this.skillLevel = skillLevel;
     }
 
 }
