+++ src/main/java/kr/co/conceptbe/common/entity/domain/TeamRecruitment.java
@@ -1,13 +1,12 @@
-package kr.co.conceptbe.common.entity;
+package kr.co.conceptbe.common.entity.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import java.util.ArrayList;
-import java.util.List;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,7 +14,7 @@ import lombok.NoArgsConstructor;
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class TeamRecruitmentCategory {
+public class TeamRecruitment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -24,12 +23,13 @@ public class TeamRecruitmentCategory {
     @Column(nullable = false)
     private String name;
 
-    @OneToMany(mappedBy = "teamRecruitmentCategory")
-    private List<TeamRecruitment> teamRecruitments = new ArrayList<>();
+    @ManyToOne
+    @JoinColumn(name = "team_recruitment_category_id")
+    private TeamRecruitmentCategory teamRecruitmentCategory;
 
-    public TeamRecruitmentCategory(Long id, String name) {
-        this.id = id;
+    public TeamRecruitment(String name, TeamRecruitmentCategory teamRecruitmentCategory) {
         this.name = name;
+        this.teamRecruitmentCategory = teamRecruitmentCategory;
     }
 
 }
