+++ src/main/java/kr/co/conceptbe/common/entity/domain/TeamRecruitmentCategory.java
@@ -1,12 +1,13 @@
-package kr.co.conceptbe.common.entity;
+package kr.co.conceptbe.common.entity.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import java.util.ArrayList;
+import java.util.List;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,7 +15,7 @@ import lombok.NoArgsConstructor;
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class TeamRecruitment {
+public class TeamRecruitmentCategory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -23,13 +24,12 @@ public class TeamRecruitment {
     @Column(nullable = false)
     private String name;
 
-    @ManyToOne
-    @JoinColumn(name = "team_recruitment_category_id")
-    private TeamRecruitmentCategory teamRecruitmentCategory;
+    @OneToMany(mappedBy = "teamRecruitmentCategory")
+    private List<TeamRecruitment> teamRecruitments = new ArrayList<>();
 
-    public TeamRecruitment(String name, TeamRecruitmentCategory teamRecruitmentCategory) {
+    public TeamRecruitmentCategory(Long id, String name) {
+        this.id = id;
         this.name = name;
-        this.teamRecruitmentCategory = teamRecruitmentCategory;
     }
 
 }
