+++ src/main/java/kr/co/conceptbe/idea/domain/IdeaBranch.java
@@ -1,4 +1,4 @@
-package kr.co.conceptbe.idea;
+package kr.co.conceptbe.idea.domain;
 
 import static lombok.AccessLevel.PROTECTED;
 
@@ -8,32 +8,34 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import kr.co.conceptbe.common.entity.base.BaseTimeEntity;
-import kr.co.conceptbe.member.Member;
+import kr.co.conceptbe.common.entity.domain.Branch;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = PROTECTED)
 @Getter
-public class IdeaLike extends BaseTimeEntity {
+public class IdeaBranch {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @ManyToOne
     @JoinColumn(name = "idea_id")
     private Idea idea;
 
-    public IdeaLike(Long id, Member member, Idea idea) {
-        this.id = id;
-        this.member = member;
+    @ManyToOne
+    @JoinColumn(name = "branch_id")
+    private Branch branch;
+
+    private IdeaBranch(Idea idea, Branch branch) {
         this.idea = idea;
+        this.branch = branch;
+    }
+
+    public static IdeaBranch of(Idea idea, Branch branch) {
+        return new IdeaBranch(idea, branch);
     }
 
 }
