+++ src/main/java/kr/co/conceptbe/idea/domain/IdeaLike.java
@@ -1,4 +1,4 @@
-package kr.co.conceptbe.idea;
+package kr.co.conceptbe.idea.domain;
 
 import static lombok.AccessLevel.PROTECTED;
 
@@ -8,34 +8,32 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import kr.co.conceptbe.common.entity.Purpose;
+import kr.co.conceptbe.common.entity.base.BaseTimeEntity;
+import kr.co.conceptbe.member.Member;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = PROTECTED)
 @Getter
-public class IdeaPurpose {
+public class IdeaLike extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne
+    @JoinColumn(name = "member_id")
+    private Member member;
+
+    @ManyToOne
     @JoinColumn(name = "idea_id")
     private Idea idea;
 
-    @ManyToOne
-    @JoinColumn(name = "purpose_id")
-    private Purpose purpose;
-
-    private IdeaPurpose(Idea idea, Purpose purpose) {
+    public IdeaLike(Long id, Member member, Idea idea) {
+        this.id = id;
+        this.member = member;
         this.idea = idea;
-        this.purpose = purpose;
-    }
-
-    public static IdeaPurpose of(Idea idea, Purpose purpose) {
-        return new IdeaPurpose(idea, purpose);
     }
 
 }
